Move inclusion path stuff to merkle package
And remove dup function bitLen()
